refactor: rename `erlab.interactive.utilities` to `erlab.interactive.utils`
