+++ src/erlab/interactive/curvefittingtool.py
@@ -8,7 +8,7 @@ import xarray as xr
 from qtpy import QtCore, QtWidgets
 
 from erlab.analysis.fit.models import MultiPeakModel
-from erlab.interactive.utilities import FittingParameterWidget, ParameterGroup
+from erlab.interactive.utils import FittingParameterWidget, ParameterGroup
 
 # EDCmultiFitting Igor procedure - 2D wave EDC fit along momentum range
